+++ yash-semantics/src/assign.rs
@@ -16,13 +16,13 @@
 
 //! Assignment.
 
+use crate::Runtime;
 use crate::expansion::AssignReadOnlyError;
 use crate::expansion::expand_value;
 use crate::xtrace::XTrace;
 use std::fmt::Write;
 use yash_env::Env;
 use yash_env::semantics::ExitStatus;
-use yash_env::system::System;
 
 #[doc(no_inline)]
 pub use crate::expansion::{Error, ErrorCause, Result};
@@ -40,7 +40,7 @@ pub use yash_syntax::syntax::Assign;
 ///
 /// If `xtrace` is `Some` instance of `XTrace`, the expanded assignment word is
 /// written to its assignments buffer.
-pub async fn perform_assignment<S: System + 'static>(
+pub async fn perform_assignment<S: Runtime + 'static>(
     env: &mut Env<S>,
     assign: &Assign,
     scope: Scope,
@@ -86,7 +86,7 @@ pub async fn perform_assignment<S: System + 'static>(
 ///
 /// If `xtrace` is `Some` instance of `XTrace`, the expanded assignment words
 /// are written to its assignments buffer.
-pub async fn perform_assignments<S: System + 'static>(
+pub async fn perform_assignments<S: Runtime + 'static>(
     env: &mut Env<S>,
     assigns: &[Assign],
     scope: Scope,