+++ yash-semantics/src/expansion/initial/param/switch.rs
@@ -20,6 +20,7 @@ use super::Env;
 use super::Error;
 use super::Phrase;
 use super::to_field;
+use crate::Runtime;
 use crate::expansion::AssignReadOnlyError;
 use crate::expansion::ErrorCause;
 use crate::expansion::attr::Origin;
@@ -27,7 +28,6 @@ use crate::expansion::attr_strip::Strip;
 use crate::expansion::expand_word;
 use crate::expansion::initial::Expand as _;
 use crate::expansion::quote_removal::skip_quotes;
-use yash_env::system::System;
 use yash_env::variable::Scope;
 use yash_env::variable::Value;
 use yash_syntax::source::Location;
@@ -193,7 +193,7 @@ fn attribute(mut phrase: Phrase) -> Phrase {
 /// As specified in the POSIX standard, this function expands the `value` and
 /// performs quote removal. The result is assigned to the variable `name` in the
 /// global scope and returned as a [`Phrase`].
-async fn assign<S: System + 'static>(
+async fn assign<S: Runtime + 'static>(
     env: &mut Env<'_, S>,
     param: &Param,
     vacancy: Vacancy,
@@ -228,7 +228,7 @@ async fn assign<S: System + 'static>(
 }
 
 /// Expands a word to be used as a vacant expansion error message.
-async fn vacant_expansion_error_message<S: System + 'static>(
+async fn vacant_expansion_error_message<S: Runtime + 'static>(
     env: &mut Env<'_, S>,
     message_word: &Word,
 ) -> Result<Option<String>, Error> {
@@ -244,7 +244,7 @@ async fn vacant_expansion_error_message<S: System + 'static>(
 }
 
 /// Constructs a vacant expansion error.
-async fn vacant_expansion_error<S: System + 'static>(
+async fn vacant_expansion_error<S: Runtime + 'static>(
     env: &mut Env<'_, S>,
     param: &Param,
     vacancy: Vacancy,
@@ -268,7 +268,7 @@ async fn vacant_expansion_error<S: System + 'static>(
 /// If this function returns `Some(_)`, that should be the result of the whole
 /// parameter expansion containing the switch. Otherwise, the parameter
 /// expansion should continue processing other modifiers.
-pub async fn apply<S: System + 'static>(
+pub async fn apply<S: Runtime + 'static>(
     env: &mut Env<'_, S>,
     switch: &Switch,
     param: &Param,