+++ yash-semantics/src/runner.rs
@@ -16,16 +16,15 @@
 
 //! Implementation of the read-eval loop
 
-use crate::Handle;
 use crate::command::Command;
 use crate::trap::run_traps_for_caught_signals;
+use crate::{Handle, Runtime};
 use std::cell::RefCell;
 use std::ops::ControlFlow::{Break, Continue};
 use yash_env::Env;
 use yash_env::semantics::Divert;
 use yash_env::semantics::ExitStatus;
 use yash_env::semantics::Result;
-use yash_env::system::System;
 use yash_syntax::parser::lex::Lexer;
 use yash_syntax::parser::{ErrorCause, Parser};
 use yash_syntax::syntax::List;
@@ -99,7 +98,7 @@ use yash_syntax::syntax::List;
 ///
 /// [`Echo`]: yash_env::input::Echo
 /// [`Input`]: yash_syntax::input::Input
-pub async fn read_eval_loop<S: System + 'static>(
+pub async fn read_eval_loop<S: Runtime + 'static>(
     env: &RefCell<&mut Env<S>>,
     lexer: &mut Lexer<'_>,
 ) -> Result {
@@ -128,7 +127,7 @@ pub async fn read_eval_loop<S: System + 'static>(
 /// [`Interrupt`]: crate::Divert::Interrupt
 /// [`Reporter`]: yash_env::input::Reporter
 /// [`IgnoreEof`]: yash_env::input::IgnoreEof
-pub async fn interactive_read_eval_loop<S: System + 'static>(
+pub async fn interactive_read_eval_loop<S: Runtime + 'static>(
     env: &RefCell<&mut Env<S>>,
     lexer: &mut Lexer<'_>,
 ) -> Result {
@@ -138,7 +137,7 @@ pub async fn interactive_read_eval_loop<S: System + 'static>(
 // The RefCell should be local to the loop, so it is safe to keep the mutable
 // borrow across await points.
 #[allow(clippy::await_holding_refcell_ref)]
-async fn read_eval_loop_impl<S: System + 'static>(
+async fn read_eval_loop_impl<S: Runtime + 'static>(
     env: &RefCell<&mut Env<S>>,
     lexer: &mut Lexer<'_>,
     is_interactive: bool,
@@ -197,7 +196,7 @@ async fn read_eval_loop_impl<S: System + 'static>(
     }
 }
 
-async fn run_command<S: System + 'static>(env: &mut Env<S>, command: &List) -> Result {
+async fn run_command<S: Runtime + 'static>(env: &mut Env<S>, command: &List) -> Result {
     run_traps_for_caught_signals(env).await?;
     env.update_all_subshell_statuses();
     command.execute(env).await