+++ yash-semantics/src/trap/exit.rs
@@ -16,10 +16,11 @@
 
 //! Running the EXIT trap
 
+use crate::Runtime;
+
 use super::run_trap;
 use std::rc::Rc;
 use yash_env::Env;
-use yash_env::system::System;
 use yash_env::trap::Action;
 use yash_env::trap::Condition;
 use yash_env::trap::Origin;
@@ -33,7 +34,7 @@ use yash_env::trap::Origin;
 /// this function and restored when finished. However, if the trap terminates
 /// with a `Break(divert)` where `divert.exit_status()` is `Some` exit status,
 /// that exit status is set to `env.exit_status`.
-pub async fn run_exit_trap<S: System + 'static>(env: &mut Env<S>) {
+pub async fn run_exit_trap<S: Runtime + 'static>(env: &mut Env<S>) {
     let Some(state) = env.traps.get_state(Condition::Exit).0 else {
         return;
     };